+++ src/api/app.py
@@ -5,7 +5,7 @@ from flask_cors import CORS
 from dotenv import load_dotenv
 from flask_migrate import Migrate
 from flasgger import Swagger
-
+from flaskr.framework.plugin.plugin_manager import enable_plugin_manager
 
 # set timezone to UTC
 # fix windows platform
@@ -33,7 +33,7 @@ def create_app() -> Flask:
 
     # init log
     init_log(app)
-
+    app = enable_plugin_manager(app)
     app.logger.info("ai-shifu-api mode: %s", app.config.get("MODE", "api"))
     # init database
     from flaskr import dao
@@ -49,7 +49,6 @@ def create_app() -> Flask:
 
     # init redis
     dao.init_redis(app)
-
     # Init LLM
     with app.app_context():
         from flaskr.api import llm  # noqa
@@ -58,7 +57,7 @@ def create_app() -> Flask:
 
     api.init_langfuse(app)
     # load plugins
-    from flaskr.util.plugin import load_plugins_from_dir
+    from flaskr.framework.plugin.load_plugin import load_plugins_from_dir
 
     load_plugins_from_dir(app, "flaskr/service/study/input")
     load_plugins_from_dir(app, "flaskr/service/study/ui")
@@ -82,6 +81,6 @@ if __name__ == "__main__":
     app.run(host="0.0.0.0", port=5800, debug=True)
 else:
     app = create_app()
-    from flaskr.util.plugin import enable_plugins
+    from flaskr.framework.plugin.enable_plugin import enable_plugins
 
     enable_plugins(app)
