+++ src/api/flaskr/route/__init__.py
@@ -22,21 +22,4 @@ def register_route(app):
     app = register_order_handler(app, prefix + "/order")
     app = register_callback_handler(app, prefix + "/callback")
     app = register_test_routes(app, prefix + "/test")
-
-    mode = app.config.get("MODE", "api")
-    app.logger.info("current run mode:{}".format(mode))
-
-    # if app.config.get("MODE", "api") == "admin":
-    #     app.logger.info("current mode is admin register admin api")
-    #     from .admin.user import register_user_route
-
-    #     app = register_user_route(app, prefix + "/user")
-    #     # register manager
-    #     from .admin.manager import register_data_manager_route
-
-    #     app = register_data_manager_route(app, prefix + "/manager")
-    #     from .admin.tools import register_manager_tools_handler
-
-    #     app = register_manager_tools_handler(app, prefix + "/tools")
-
     return app
